fix: improved code quality according to CodeClimate
