+++ tests/end2end/helpers/E2EApp.js
@@ -73,7 +73,7 @@ export class E2EApp {
                 let userWantsLdap = tab === '#tab_ldap';
 
                 return (tabIsStandard && userWantsStandard) || (tabIsLdap && userWantsLdap);
-            }, 5000);
+            }, 7500, 'The login screen could not been loaded in time');
 
             let tabIsStandard = browser.isExisting('#at-field-username_and_email');
             let tabIsLdap = browser.isExisting('#id_ldapUsername');
@@ -93,7 +93,7 @@ export class E2EApp {
             browser.waitUntil(_ => {
                 const userMenuExists = browser.isExisting('#navbar-usermenu');
                 return userMenuExists || E2EApp.loginFailed();
-            }, 20000);
+            }, 20000, 'The login could not been processed in time');
 
             if (E2EApp.loginFailed()) {
                 throw new Error ("Unknown user or wrong password.");
